pendiente la correccion de *Cita*Permisohistorial*Diagnostico*Recomendación*Receta*Resenia

+++ src/main/java/com/habimed/habimedWebService/consultorio/domain/service/ConsultorioServiceImpl.java
@@ -1,6 +1,5 @@
 package com.habimed.habimedWebService.consultorio.domain.service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,43 +18,32 @@ public class ConsultorioServiceImpl implements ConsultorioService {
     private ConsultorioServiceImpl(ConsultorioRepository consultorioRepository) {
         this.consultorioRepository = consultorioRepository;
     }
-    // Aquí puedes implementar los métodos de ConsultorioService
-    // Por ejemplo, podrías tener un repositorio inyectado para acceder a la base de datos
 
     @Override
     public List<ConsultorioDTO> getAllConsultorios(ConsultorioRequest request) {
         // Implementación para obtener todos los consultorios
-        List<ConsultorioDTO> consultorios = consultorioRepository.getAllConsultorios(request);
-
-        return consultorios;
+        return consultorioRepository.getAllConsultorios(request);
     }
 
     @Override
     public ConsultorioDTO getConsultorioById(Integer id) {
         // Implementación para obtener un consultorio por ID
-        ConsultorioDTO consultorio = consultorioRepository.getConsultorio(id);
-        return consultorio;
+        return consultorioRepository.getConsultorio(id);
     }
 
     @Override
     public Integer setConsultorio(ConsultorioRequest request) {
         // Implementación para crear un nuevo consultorio
-        Integer idConsultorio = consultorioRepository.setConsultorio(request);
-        return idConsultorio;
-    }
-
-    @Override
-    public boolean updateConsultorio(ConsultorioRequest request) {
-        // Implementación para actualizar un consultorio existente
-        boolean isUpdated = consultorioRepository.updateConsultorio(request);
-        return isUpdated;
+        return consultorioRepository.setConsultorio(request);
     }
 
     @Override
     public boolean deleteConsultorio(Integer id) {
-        // Implementación para eliminar un consultorio por ID
-        boolean isDeleted = consultorioRepository.deleteConsultorio(id);
-        return isDeleted;
+        if( id == null ){
+            return false;
+        }else{
+            return consultorioRepository.deleteConsultorio(id);
+        }
     }
     
 }
